Global search: compute the last modified time
Moodle does not document it, but it seems to use a modifiedtime
field in the DB, which mod_etherpad does not keep up to date.
This commit checks in real-time if any pad has changed.

A slighlty faster solution would be to update this modifiedtime field
through a cron. It would introduce latency in the search indexing,
but might fix other problems caused by this out-of-date field.